+++ cli/k8s_client/yaml_factory.go
@@ -248,6 +248,9 @@ spec:
           initialDelaySeconds: 120
           periodSeconds: 120
           timeoutSeconds: 90
+      nodeSelector:
+        beta.kubernetes.io/os: linux
+        beta.kubernetes.io/arch: amd64
 `
 
 func GetCSIServiceYAML(label string) string {
@@ -406,6 +409,9 @@ spec:
         volumeMounts:
         - name: socket-dir
           mountPath: /var/lib/csi/sockets/pluginproxy/
+      nodeSelector:
+        beta.kubernetes.io/os: linux
+        beta.kubernetes.io/arch: amd64
       volumes:
       - name: socket-dir
         emptyDir:
@@ -508,6 +514,9 @@ spec:
         volumeMounts:
         - name: socket-dir
           mountPath: /var/lib/csi/sockets/pluginproxy/
+      nodeSelector:
+        kubernetes.io/os: linux
+        kubernetes.io/arch: amd64
       volumes:
       - name: socket-dir
         emptyDir:
@@ -611,6 +620,9 @@ spec:
         volumeMounts:
         - name: socket-dir
           mountPath: /var/lib/csi/sockets/pluginproxy/
+      nodeSelector:
+        kubernetes.io/os: linux
+        kubernetes.io/arch: amd64
       volumes:
       - name: socket-dir
         emptyDir:
@@ -728,6 +740,9 @@ spec:
           mountPath: /plugin
         - name: registration-dir
           mountPath: /registration
+      nodeSelector:
+        beta.kubernetes.io/os: linux
+        beta.kubernetes.io/arch: amd64
       volumes:
       - name: plugin-dir
         hostPath:
@@ -847,6 +862,9 @@ spec:
           mountPath: /plugin
         - name: registration-dir
           mountPath: /registration
+      nodeSelector:
+        kubernetes.io/os: linux
+        kubernetes.io/arch: amd64
       volumes:
       - name: plugin-dir
         hostPath:
@@ -1067,6 +1085,9 @@ spec:
     volumeMounts:
     - name: etcd-vol
       mountPath: /var/etcd/data
+  nodeSelector:
+    beta.kubernetes.io/os: linux
+    beta.kubernetes.io/arch: amd64
   volumes:
   - name: etcd-vol
     persistentVolumeClaim:
@@ -1101,6 +1122,9 @@ spec:
     - name: setup-dir
       mountPath: /setup
   restartPolicy: Never
+  nodeSelector:
+    beta.kubernetes.io/os: linux
+    beta.kubernetes.io/arch: amd64
   volumes:
   - name: setup-dir
     configMap:
@@ -1131,6 +1155,9 @@ spec:
     image: {TRIDENT_IMAGE}
     workingDir: /
     command: {COMMAND}
+  nodeSelector:
+    beta.kubernetes.io/os: linux
+    beta.kubernetes.io/arch: amd64
   restartPolicy: Never
 `